refactor: cli cmd node running
Extract the load and run logic from the cli.

Have the main.go file only deal the the 'CLI' environment,
then call a public function for the reset.
